Prevent errors if no picture url

+++ server/app/services/question-pictures-deletion/question-pictures-deletion.service.ts
@@ -34,12 +34,12 @@ export class QuestionPicturesDeletionService {
     }
 
     async deleteNonUsedPicture(pictureUrl: string, games: Game[]) {
+        if (!pictureUrl) return;
         const nGamesWithSameImage = this.countGamesSamePicture(pictureUrl, games);
         const nBankQuestionsWithSameImage = await this.questionService.countQuestionsSamePicture(pictureUrl);
         console.log(`nGames: ${nGamesWithSameImage}`);
         console.log(`nBankQuestions: ${nBankQuestionsWithSameImage}`);
         if (nGamesWithSameImage + nBankQuestionsWithSameImage === 0) {
-            console.log(pictureUrl);
             await this.firebaseRepositoryService.deleteImage(pictureUrl);
         }
     }
